+++ phpBB2/modcp.php
@@ -409,86 +409,22 @@ switch( $mode )
 			if ($new_forum_id <= 0 ) message_die(GENERAL_MESSAGE, $lang['Forum_not_exist']);
 //-- fin mod : categories hierarchy ----------------------------------------------------------------
 			$old_forum_id = $forum_id;
-			$sql = 'SELECT forum_id FROM ' . FORUMS_TABLE . '
-				WHERE forum_id = ' . $new_forum_id;
-			if ( !($result = $db->sql_query($sql)) )
+			$topics = !empty($topic_id_list) ? $topic_id_list : array($topic_id);
+			require_once $phpbb_root_path . 'includes/functions_topic_move.' . $phpEx;
+			try
 			{
-				message_die(GENERAL_ERROR, 'Could not select from forums table', '', __LINE__, __FILE__, $sql);
+				$moved = phpbb_move_topics($db, $old_forum_id, $new_forum_id, $topics, isset($_POST['move_leave_shadow']));
 			}
-
-			if (!$db->sql_fetchrow($result))
+			catch (PhpbbTopicMoveException $error)
 			{
-				message_die(GENERAL_MESSAGE, 'New forum does not exist');
+				message_die(GENERAL_MESSAGE, $error->getMessage());
 			}
-
-			$db->sql_freeresult($result); 
-			if ( $new_forum_id != $old_forum_id )
+			if ($moved)
 			{
-				$topics = !empty($topic_id_list) ? $topic_id_list : array($topic_id);
-
-				$topic_list = '';
-				for($i = 0; $i < count($topics); $i++)
-				{
-					$topic_list .= ( ( $topic_list != '' ) ? ', ' : '' ) . intval($topics[$i]);
-				}
-
-				$sql = "SELECT * 
-					FROM " . TOPICS_TABLE . " 
-					WHERE topic_id IN ($topic_list)
-						AND forum_id = $old_forum_id
-						AND topic_status <> " . TOPIC_MOVED;
-				if ( !($result = $db->sql_query($sql, BEGIN_TRANSACTION)) )
-				{
-					message_die(GENERAL_ERROR, 'Could not select from topic table', '', __LINE__, __FILE__, $sql);
-				}
-
-				$row = $db->sql_fetchrowset($result);
-				$db->sql_freeresult($result);
-
-				for($i = 0; $i < count($row); $i++)
-				{
-					$topic_id = $row[$i]['topic_id'];
-					
-					if ( isset($_POST['move_leave_shadow']) )
-					{
-						// Insert topic in the old forum that indicates that the forum has moved.
-						$sql = "INSERT INTO " . TOPICS_TABLE . " (forum_id, topic_title, topic_poster, topic_time, topic_status, topic_type, topic_vote, topic_views, topic_replies, topic_first_post_id, topic_last_post_id, topic_moved_id)
-							VALUES ($old_forum_id, '" . addslashes(str_replace("\'", "''", $row[$i]['topic_title'])) . "', '" . str_replace("\'", "''", $row[$i]['topic_poster']) . "', " . $row[$i]['topic_time'] . ", " . TOPIC_MOVED . ", " . POST_NORMAL . ", " . $row[$i]['topic_vote'] . ", " . $row[$i]['topic_views'] . ", " . $row[$i]['topic_replies'] . ", " . $row[$i]['topic_first_post_id'] . ", " . $row[$i]['topic_last_post_id'] . ", $topic_id)";
-						if ( !$db->sql_query($sql) )
-						{
-							message_die(GENERAL_ERROR, 'Could not insert shadow topic', '', __LINE__, __FILE__, $sql);
-						}
-					}
-
-					$sql = "UPDATE " . TOPICS_TABLE . " 
-						SET forum_id = $new_forum_id  
-						WHERE topic_id = $topic_id";
-					if ( !$db->sql_query($sql) )
-					{
-						message_die(GENERAL_ERROR, 'Could not update old topic', '', __LINE__, __FILE__, $sql);
-					}
-
-					$sql = "UPDATE " . POSTS_TABLE . " 
-						SET forum_id = $new_forum_id 
-						WHERE topic_id = $topic_id";
-					if ( !$db->sql_query($sql) )
-					{
-						message_die(GENERAL_ERROR, 'Could not update post topic ids', '', __LINE__, __FILE__, $sql);
-					}
-				}
-
-				// Sync the forum indexes
-				sync('forum', $new_forum_id);
-				sync('forum', $old_forum_id);
-				log_action('move', $topic_list, $userdata['user_id'], $userdata['username']);
-
-				$message = $lang['Topics_Moved'] . '<br /><br />';
-
+				board_stats();
+				cache_tree(true);
 			}
-			else
-			{
-				$message = $lang['No_Topics_Moved'] . '<br /><br />';
-			}
+			$message = $lang[$moved ? 'Topics_Moved' : 'No_Topics_Moved'] . '<br /><br />';
 
 			if ( !empty($topic_id) )
 			{